requirements.txt
# --------- pytorch --------- #
torch==1.13.0
torchvision==0.14.0
pytorch_lightning==1.8.1
torchmetrics==0.10.0
torchaudio==0.13.0

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0

# --------- audio --------- #
audiofile==1.1.0
pydub==0.25.1
nnAudio==0.3.2

# --------- loggers --------- #
wandb

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
mega.py==1.0.8
progressbar